+++ tests/pyxelresttest_duplicated_service_configurations.py
@@ -14,11 +14,10 @@ class PyxelRestDuplicatedServiceConfigurationTest(unittest.TestCase):
         serviceshandler.stop_services()
 
     def test_without_service_configuration_file(self):
-        try:
+        with self.assertRaises(Exception) as cm:
             loader.load('pyxelresttest_duplicated_service_configuration.ini')
-        except Exception as e:
-            self.assertEqual(str(e), "While reading from '{0}' [line  5]: section 'usual_parameters' already exists".format(
-                pyxelrest.SERVICES_CONFIGURATION_FILE_PATH.replace('\\', '\\\\')))
+        self.assertEqual(cm.exception.args[0], "While reading from '{0}' [line  5]: section 'usual_parameters' already exists".format(
+            pyxelrest.SERVICES_CONFIGURATION_FILE_PATH.replace('\\', '\\\\')))
 
 
 if __name__ == '__main__':